Change WroDSLHandler to always return a new instance of the DSL object
GroovyModelParser assumes that it is always passed a new object, and it doesn't work right if the same object is passed a second time. The problem is that WroModelDelegate will execute a second time on the same object, causing the returned wroModel to be empty (there won't be any groups in it).